fix issue with the transit api not liking non integers

+++ lib/api/TransitManager.dart
@@ -94,7 +94,7 @@ class TransitManager {
     return result;
   }
 
-  String genStopLocationURL(double long, double lat, double distance, bool walking){
+  String genStopLocationURL(double long, double lat, int distance, bool walking){
     URLGenerator url = URLGenerator(url: sprintf(apiUrl, ["stops"]));
     url.addParam("lat", lat.toString());
     url.addParam("lon", long.toString());
@@ -103,7 +103,7 @@ class TransitManager {
     return url.toString();
   }
 
-  Future<List<BusStop>> genStopLocations(double long, double lat, double distance, bool walking) async {
+  Future<List<BusStop>> genStopLocations(double long, double lat, int distance, bool walking) async {
     List<BusStop> busStops = [];
     BusStop busStop;
